Modularize eval/speech routes and tool execution
Refactors `src/api/routes.ts` by moving eval and speech endpoints into dedicated route modules (`routes/evals.ts` and `routes/speech.ts`) and wiring them back through route spreads. Also extracts tool-call execution flow from `agent-provider-runtime` into a new shared `agent-tool-execution` module, preserving hook/status behavior and adding regression tests for unknown/disabled tools and successful execution state updates.
